api/requirements.txt
flask==3.1.0
pillow==11.1.0
tensorflow==2.18.0
numpy==1.26.4